test(openapi): extract findRequestByName/findFolderByName into shared helper
