+++ packages/bruno-converters/tests/openapi/openapi-to-bruno/openapi-tags.spec.js
@@ -1,39 +1,6 @@
 import { describe, it, expect } from '@jest/globals';
 import openApiToBruno from '../../../src/openapi/openapi-to-bruno';
-
-/**
- * Helper function to find a request by name in the collection.
- * Searches recursively through folders since requests with tags
- * are grouped into folders.
- */
-const findRequestByName = (items, name) => {
-  for (const item of items) {
-    if (item.type === 'http-request' && item.name === name) {
-      return item;
-    }
-    if (item.type === 'folder' && item.items) {
-      const found = findRequestByName(item.items, name);
-      if (found) return found;
-    }
-  }
-  return undefined;
-};
-
-/**
- * Helper function to find a folder by name in the collection.
- */
-const findFolderByName = (items, name) => {
-  for (const item of items) {
-    if (item.type === 'folder' && item.name === name) {
-      return item;
-    }
-    if (item.type === 'folder' && item.items) {
-      const found = findFolderByName(item.items, name);
-      if (found) return found;
-    }
-  }
-  return undefined;
-};
+import { findRequestByName, findFolderByName } from '../../common/find-items';
 
 describe('OpenAPI Import - Tag Sanitization', () => {
   it('should replace spaces with underscores in tags', () => {
