+++ packages/bruno-converters/tests/openapi/swagger2-to-bruno/swagger2-docs.spec.js
@@ -1,27 +1,6 @@
 import { describe, it, expect } from '@jest/globals';
 import swagger2ToBruno from '../../../src/openapi/swagger2-to-bruno';
-
-const findRequestByName = (items, name) => {
-  for (const item of items) {
-    if (item.type === 'http-request' && item.name === name) return item;
-    if (item.type === 'folder' && item.items) {
-      const found = findRequestByName(item.items, name);
-      if (found) return found;
-    }
-  }
-  return undefined;
-};
-
-const findFolderByName = (items, name) => {
-  for (const item of items) {
-    if (item.type === 'folder' && item.name === name) return item;
-    if (item.type === 'folder' && item.items) {
-      const found = findFolderByName(item.items, name);
-      if (found) return found;
-    }
-  }
-  return undefined;
-};
+import { findRequestByName, findFolderByName } from '../../common/find-items';
 
 const buildSpec = (overrides = {}) => ({
   swagger: '2.0',
